Removing files directory
This only contained a certificates directory, which in line with changes to the base flavour, should only be present in flavours where web-servers are assumed, i.e. in web flavours.

As this flavour would not make this assumption the certificates directory will be removed, and since this will leave the files directory empty, the files directory will be removed as well.